+++ src/main/java/com/karel/game/gridobjects/gridentities/zombies/hornetneck/HornetNeckZombie.java
@@ -17,7 +17,7 @@ public class HornetNeckZombie extends Zombie
     private static ZombieClass[] classes = new ZombieClass[]{ZombieClass.spawner, ZombieClass.assault};
     private static final int gunReloadTime = 5;         // The minimum delay between firing the gun.
 
-    private int reloadDelayCount;               // How long ago we fired the gun the last time.
+    private double reloadDelayCount;               // How long ago we fired the gun the last time.
 
     public String getStaticTextureURL(){return "hornetnestzareln.png";}
     private int bees = 20;
